+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added a `debug test-sentry` command to test Sentry connection
 - Manual interaction mode (using admin-enable command) to use the bot on specific pull requests only
 - Optional history tracking
+- Support for more admin commands (`set-default-needed-reviewers`, `set-default-merge-strategy`, `set-default-pr-title-regex`, `set-needed-reviewers`)
 
 ### Fixed